+++ internal/provider/resource_azure_subscription.go
@@ -108,6 +108,10 @@ are used when specifying the feature set.
   * ´CUSTOMER_MANAGED_BASIC´ - Represents the permissions required to enable
     customer-managed Exocompute feature.
 
+´SERVERS_AND_APPS´
+  * ´CLOUD_CLUSTER_ES´ - Represents the basic set of permissions required to onboard the
+    feature.
+
 ~> **Note:** Even though the ´resource_group_name´ and the
    ´resource_group_region´ fields are marked as optional you should always
    specify them. They are marked as optional to simplify the migration of
@@ -222,6 +226,7 @@ func resourceAzureSubscription() *schema.Resource {
 					keyExocompute,
 					keySQLDBProtection,
 					keySQLMIProtection,
+					keyServersAndApps,
 				},
 				Description: "Enable the RSC Cloud Native Archival feature for the Azure subscription. Provides " +
 					"archival of data from workloads for disaster recovery and long-term retention.",
@@ -384,6 +389,7 @@ func resourceAzureSubscription() *schema.Resource {
 					keyExocompute,
 					keySQLDBProtection,
 					keySQLMIProtection,
+					keyServersAndApps,
 				},
 				Description: "Enable the RSC Cloud Native Protection feature for Azure Blob Storage. Provides " +
 					"protection for Azure Blob Storage through the rules and policies of SLA Domains.",
@@ -473,6 +479,7 @@ func resourceAzureSubscription() *schema.Resource {
 					keyExocompute,
 					keySQLDBProtection,
 					keySQLMIProtection,
+					keyServersAndApps,
 				},
 				Description: "Enable the RSC Cloud Native Protection feature for the Azure subscription. Provides " +
 					"protection for Azure virtual machines and managed disks through the rules and policies of SLA " +
@@ -567,9 +574,98 @@ func resourceAzureSubscription() *schema.Resource {
 					keyCloudNativeProtection,
 					keySQLDBProtection,
 					keySQLMIProtection,
+					keyServersAndApps,
 				},
 				Description: "Enable the RSC Exocompute feature for the Azure subscription. Provides snapshot " +
 					"indexing, file recovery, storage tiering, and application-consistent protection of Azure objects.",
+			},
+			keyServersAndApps: {
+				Type: schema.TypeList,
+				Elem: &schema.Resource{
+					Schema: map[string]*schema.Schema{
+						keyPermissionGroups: {
+							Type: schema.TypeSet,
+							Elem: &schema.Schema{
+								Type: schema.TypeString,
+								ValidateFunc: validation.StringInSlice([]string{
+									"CLOUD_CLUSTER_ES",
+								}, false),
+							},
+							Optional: true,
+							Description: "Permission groups to assign to the Cloud Cluster feature. " +
+								"Possible values are `CLOUD_CLUSTER_ES`.",
+						},
+						keyPermissions: {
+							Type:     schema.TypeString,
+							Optional: true,
+							Description: "Permissions updated signal. When this field changes, the provider will notify " +
+								"RSC that the permissions for the feature has been updated. Use this field with the " +
+								"`polaris_azure_permissions` data source.",
+							ValidateFunc: validation.StringIsNotWhiteSpace,
+						},
+						keyRegions: {
+							Type: schema.TypeSet,
+							Elem: &schema.Schema{
+								Type: schema.TypeString,
+							},
+							MinItems: 1,
+							Required: true,
+							Description: "Azure regions to enable the Cloud Cluster feature in. Should be specified " +
+								"in the standard Azure style, e.g. `eastus`.",
+						},
+						keyResourceGroupName: {
+							Type:     schema.TypeString,
+							Optional: true,
+							RequiredWith: []string{
+								keyServersAndApps + ".0." + keyResourceGroupRegion,
+							},
+							Description: "Name of the Azure resource group where RSC places all resources created by " +
+								"the feature. RSC assumes the resource group already exists. Changing this forces the " +
+								"RSC feature to be re-onboarded.",
+							ValidateFunc: validation.StringIsNotWhiteSpace,
+						},
+						keyResourceGroupRegion: {
+							Type:     schema.TypeString,
+							Optional: true,
+							RequiredWith: []string{
+								keyServersAndApps + ".0." + keyResourceGroupName,
+							},
+							Description: "Region of the Azure resource group. Should be specified in the standard " +
+								"Azure style, e.g. `eastus`. Changing this forces the RSC feature to be re-onboarded.",
+							ValidateFunc: validation.StringInSlice(gqlregion.AllRegionNames(), false),
+						},
+						keyResourceGroupTags: {
+							Type: schema.TypeMap,
+							Elem: &schema.Schema{
+								Type: schema.TypeString,
+							},
+							Optional: true,
+							RequiredWith: []string{
+								keyServersAndApps + ".0." + keyResourceGroupName,
+								keyServersAndApps + ".0." + keyResourceGroupRegion,
+							},
+							Description: "Tags to add to the Azure resource group. Changing this forces the RSC feature " +
+								"to be re-onboarded.",
+						},
+						keyStatus: {
+							Type:        schema.TypeString,
+							Computed:    true,
+							Description: "Status of the Cloud Cluster feature.",
+						},
+					},
+				},
+				MaxItems: 1,
+				Optional: true,
+				AtLeastOneOf: []string{
+					keyCloudNativeArchival,
+					keyCloudNativeBlobProtection,
+					keyCloudNativeProtection,
+					keyExocompute,
+					keySQLDBProtection,
+					keySQLMIProtection,
+				},
+				Description: "Enable the RSC Cloud Cluster feature for the Azure subscription. Provides " +
+					"ability to deploy Rubrik Cloud Data Management (CDM) clusters in Azure.",
 			},
 			keySQLDBProtection: {
 				Type: schema.TypeList,
@@ -654,6 +750,7 @@ func resourceAzureSubscription() *schema.Resource {
 					keyCloudNativeProtection,
 					keyExocompute,
 					keySQLMIProtection,
+					keyServersAndApps,
 				},
 				Description: "Enable the RSC SQL DB Protection feature for the Azure subscription. Provides " +
 					"centralized database backup management and recovery in an Azure SQL Database deployment.",
@@ -707,6 +804,7 @@ func resourceAzureSubscription() *schema.Resource {
 					keyCloudNativeProtection,
 					keyExocompute,
 					keySQLDBProtection,
+					keyServersAndApps,
 				},
 				Description: "Enable the RSC SQL MI Protection feature for the Azure subscription. Provides " +
 					"centralized database backup management and recovery for an Azure SQL Managed Instance deployment.",
@@ -1115,6 +1213,13 @@ var azureKeyFeatureMap = map[string]orderedFeature{
 		orderRemove:      306,
 		orderSplitAdd:    213,
 		orderSplitRemove: 212,
+	},
+	keyServersAndApps: {
+		feature:          core.FeatureServerAndApps,
+		orderAdd:         107,
+		orderRemove:      307,
+		orderSplitAdd:    215,
+		orderSplitRemove: 214,
 	},
 }