+++ docs/guides/changelog.md
@@ -28,7 +28,7 @@ page_title: "Changelog"
   feature and permissions groups to look up the GCP permissions for.
 * Add `conditions` field to the `polaris_gcp_permissions` data source. The `conditions` field holds the GCP conditions
   for the `with_conditions` GCP permissions.
-* Add `services` field to the `polaris_gcp_permissions` data source. The `servcices` field holds the GCP services
+* Add `services` field to the `polaris_gcp_permissions` data source. The `services` field holds the GCP services
   required for the RSC feature and permission groups.
 * Add `with_conditions` field to the `polaris_gcp_permissions` data source. The `with_conditions` field holds the GCP
   permissions with conditions required for the RSC feature and permission groups.
@@ -47,6 +47,11 @@ page_title: "Changelog"
 * The `project`, `project_name`, and `project_number` fields of the `polaris_gcp_project` resource are now marked as
   required. Note, this is a breaking change if they are not specified in a Terraform configuration. Use the
   `terraform state show <resource-address>` command to look up the values to add to the configuration.
+* Adds `servers_and_apps` field to the `polaris_azure_subscription` resource. The `servers_and_apps` field is used to
+  specify the permission groups for the `SERVERS_AND_APPS` feature which is required for provisioning an Azure cloud
+  cluster.
+* Adds a new resource `polaris_azure_cloud_cluster`. The `polaris_azure_cloud_cluster` resource creates an Azure cloud
+  cluster using RSC. [[docs](../resources/azure_cloud_cluster.md)]
 
 ## v1.2.1
 * Update Go version to 1.24.8.